==> 15. Deepfake tweets detection/Project-1/notebooks/04_bert_word_embeddings_ml.ipynb <==
# Load packages

import os

import numpy as np
import pandas as pd
import transformers
import torch

device = "cuda:0" if torch.cuda.is_available() else "cpu"

## Load data

DATA_PATH = os.path.join('..', 'data', 'preprocessed_url_simple')
train = pd.read_csv(os.path.join(DATA_PATH, 'train.csv'))
validation = pd.read_csv(os.path.join(DATA_PATH, 'validation.csv'))
test = pd.read_csv(os.path.join(DATA_PATH, 'test.csv'))
train.head()

train.head()

spec = "bert-base-cased"
tokenizer = transformers.BertTokenizer.from_pretrained(spec, max_length=512)
model = transformers.BertModel.from_pretrained(spec).to(device)

token_test = tokenizer.encode(train['text'][0], truncation=True, max_length=512, return_tensors='pt').to(device)
ans = model(token_test)

ans[0].shape

def add_bert_embeddings(df):
    """
    Add BERT embeddings to dataframe
    :param df: dataframe with column 'text'
    :return: dataframe with additional column 'bert_embeddings'
    """

    def transform(text):
        token = tokenizer.encode(text, truncation=True, max_length=512, return_tensors='pt').to(device)
        ans = model(token)
        vec = np.array(ans[0].cpu().detach().numpy())
        vec = vec.reshape(vec.shape[1], vec.shape[2])
        vec = np.mean(vec, axis=0)
        return vec

    df['bert_embeddings'] = df['text'].apply(lambda x: transform(x))
    return df


train = add_bert_embeddings(train)
train.head()

train['bert_embeddings'][0].shape

validation = add_bert_embeddings(validation)
test = add_bert_embeddings(test)

## Save data

OUTPUT_PATH = os.path.join('..', 'data', 'bert_embeddings')
os.makedirs(OUTPUT_PATH, exist_ok=True)
train.to_pickle(os.path.join(OUTPUT_PATH, 'train.pkl'))
validation.to_pickle(os.path.join(OUTPUT_PATH, 'validation.pkl'))
test.to_pickle(os.path.join(OUTPUT_PATH, 'test.pkl'))
